Lots of rafactoring

+++ app/models/repository.rb
@@ -1,56 +1,52 @@
 require 'rugged'
 require 'fileutils'
+require_relative 'repository/diff'
 
 # Provides the ability to clone a repository to a standard location and
 # provide access to that repository.
 class Repository
   class << self
-    def clone(repositories_path, pull_request, build_id)
+    def clone(repos_dir, task)
       repo = Rugged::Repository.clone_at(
-        pull_request.clone_url,
-        clone_directory(repositories_path, pull_request.slug, build_id),
-        checkout_branch: pull_request.branch
+        task.clone_url,
+        File.join(repos_dir, task.slug)
       )
+      repo.checkout(task.head_sha, strategy: :force)
 
-      yield (repository = new(repo, pull_request.slug))
+      repository = new(repo)
 
-      repository.destroy!
-    end
+      if block_given?
+        yield repository
 
-  private
+        repository.destroy!
+      end
 
-    def clone_directory(repositories_path, slug, build_id)
-      File.join(repositories_path, slug, build_id.to_s)
+      repository
     end
   end
 
-  attr_reader :slug
-
-  def initialize(repo, slug)
+  def initialize(repo)
     @repo = repo
-    @slug = slug
   end
 
-  def modified_files(pull_request)
-    diff = repo.diff(pull_request.base_sha, pull_request.head_sha)
-
-    ModifiedFiles.new(diff, local_path)
+  def diff(base_sha, head_sha)
+    Repository::Diff.new(repo, base_sha, head_sha)
   end
 
-  def branch
-    repo.head.name.sub('refs/heads/', '')
+  def modified_files(base_sha, head_sha)
+    diff(base_sha, head_sha).modified_files
   end
 
-  def local_path
-    repo.workdir
+  def head_sha
+    repo.head.target.oid
   end
 
   def destroy!
-    FileUtils.rm_rf(local_path)
+    FileUtils.rm_rf(repo.workdir)
   end
 
   def inspect
-    "<Repository: #{slug}>"
+    "<Repository: #{repo.workdir}>"
   end
 
 private
